requirements.txt
streamlit==1.32.2
pandas==2.2.1
scikit-learn==1.4.1.post1
numpy==1.26.4
matplotlib==3.8.3
seaborn==0.13.2
google.generativeai==0.6.0
openpyxl==3.1.4
langchain-core==0.2.9
langchain-google-genai==0.7.0
python-dotenv==1.0.1


